refactor: address SQL client review feedback
