+++ rust/lancedb/src/remote/sql.rs
@@ -18,15 +18,17 @@ use arrow_flight::{
 };
 use arrow_schema::{Schema, SchemaRef};
 use futures::TryStreamExt;
+use http::header::{HeaderMap, HeaderName, HeaderValue};
 use prost::Message;
-use tokio::sync::{Mutex, Notify, OnceCell, OwnedSemaphorePermit, Semaphore, mpsc};
+use tokio::sync::{Mutex, Notify, OnceCell, mpsc};
 use tonic::transport::{Certificate, Channel, ClientTlsConfig, Endpoint, Identity};
+use uuid::Uuid;
 
 use crate::arrow::{SendableRecordBatchStream, SimpleRecordBatchStream};
 use crate::error::{Error, Result};
 use crate::remote::client::{ClientConfig, TlsConfig};
 use crate::remote::retry::ResolvedRetryConfig;
-use crate::sql::{Query, QueryDescription, QueryHandle};
+use crate::sql::{Query, QueryDescription, QueryHandle, QueryStatus};
 
 const DEFAULT_SQL_PORT: u16 = 10025;
 const DEFAULT_SQL_TLS_PORT: u16 = 10026;
@@ -35,10 +37,8 @@ const DEFAULT_READ_TIMEOUT: Duration = Duration::from_secs(300);
 const STATUS_POLL_TIMEOUT: Duration = Duration::from_secs(1);
 const MIN_POLL_INTERVAL: Duration = Duration::from_millis(50);
 const MAX_SQL_MESSAGE_SIZE: usize = 1024 * 1024 * 1024;
-const QUERY_CACHE_CAPACITY: u64 = 10_000;
 const TERMINAL_QUERY_RETENTION: Duration = Duration::from_secs(300);
 const ABANDONED_QUERY_RETENTION: Duration = Duration::from_secs(24 * 60 * 60);
-const QUERY_ID_PREFIX: &str = "lq1_";
 
 #[derive(Clone)]
 pub(super) struct SqlClient {
@@ -57,6 +57,8 @@ struct SqlClientInner {
 }
 
 struct SqlConnection {
+    // FlightClient does not expose its transport. Cancellation retains the channel so it can
+    // install a per-call interceptor that records whether a request was dispatched.
     channel: Channel,
     client: FlightServiceClient<Channel>,
 }
@@ -176,7 +178,7 @@ impl SqlClient {
                 client_config,
                 client: Arc::new(OnceCell::new()),
             }),
-            queries: Arc::new(QueryRegistry::new(QUERY_CACHE_CAPACITY)),
+            queries: Arc::new(QueryRegistry::new()),
         }
     }
 
@@ -188,27 +190,26 @@ impl SqlClient {
         let timeout = self.inner.overall_timeout()?;
         with_overall_timeout(timeout, "SQL query submission", async {
             validate_namespace_path(default_namespace_path)?;
-            let permit = self.queries.reserve()?;
             let command = CommandStatementQuery {
                 query: query.to_string(),
                 transaction_id: None,
             };
             let descriptor = FlightDescriptor::new_cmd(command.as_any().encode_to_vec());
             let poll_info = self.inner.poll(descriptor, default_namespace_path).await?;
-            let query_id = format!("{QUERY_ID_PREFIX}{}", uuid::Uuid::new_v4().simple());
+            let query_id = Uuid::now_v7();
             let query = Arc::new(RemoteQuery::new(
-                query_id.clone(),
+                query_id,
                 self.inner.clone(),
                 default_namespace_path.to_vec(),
                 poll_info,
             )?);
-            self.queries.insert(query_id, query.clone(), permit);
+            self.queries.insert(query_id, query.clone());
             Ok(Query::new(Arc::new(RemoteQueryHandle::new(query))))
         })
         .await
     }
 
-    pub(super) async fn describe(&self, query_id: &str) -> Result<QueryDescription> {
+    pub(super) async fn describe(&self, query_id: Uuid) -> Result<QueryDescription> {
         let query = self
             .queries
             .get(query_id)
@@ -449,8 +450,8 @@ impl SqlClientInner {
         &self,
         default_namespace_path: &[String],
         request_id: &str,
-    ) -> Result<HashMap<String, String>> {
-        let mut headers = HashMap::new();
+    ) -> Result<HeaderMap> {
+        let mut headers = HeaderMap::new();
         merge_headers(&mut headers, &self.client_config.extra_headers)?;
         if let Some(provider) = &self.client_config.header_provider {
             merge_headers(&mut headers, &provider.get_headers().await?)?;
@@ -490,61 +491,25 @@ impl SqlClientInner {
     }
 }
 
-struct RegisteredQuery {
-    query: Arc<RemoteQuery>,
-    _permit: OwnedSemaphorePermit,
-}
-
 struct QueryRegistry {
-    queries: StdMutex<HashMap<String, RegisteredQuery>>,
-    capacity: Arc<Semaphore>,
-    max_capacity: u64,
+    queries: StdMutex<HashMap<Uuid, Arc<RemoteQuery>>>,
 }
 
 impl QueryRegistry {
-    fn new(capacity: u64) -> Self {
+    fn new() -> Self {
         Self {
             queries: StdMutex::new(HashMap::new()),
-            capacity: Arc::new(Semaphore::new(capacity as usize)),
-            max_capacity: capacity,
         }
     }
 
-    fn reserve(&self) -> Result<OwnedSemaphorePermit> {
+    fn insert(&self, id: Uuid, query: Arc<RemoteQuery>) {
         self.remove_expired();
-        if let Ok(permit) = self.capacity.clone().try_acquire_owned() {
-            return Ok(permit);
-        }
-        if let Some(permit) = self
-            .remove_oldest_terminal()
-            .or_else(|| self.remove_oldest_abandoned())
-        {
-            return Ok(permit);
-        }
-        self.capacity
-            .clone()
-            .try_acquire_owned()
-            .map_err(|_| self.capacity_error())
+        self.queries.lock().unwrap().insert(id, query);
     }
 
-    fn insert(&self, id: String, query: Arc<RemoteQuery>, permit: OwnedSemaphorePermit) {
-        self.queries.lock().unwrap().insert(
-            id,
-            RegisteredQuery {
-                query,
-                _permit: permit,
-            },
-        );
-    }
-
-    fn get(&self, id: &str) -> Option<Arc<RemoteQuery>> {
+    fn get(&self, id: Uuid) -> Option<Arc<RemoteQuery>> {
         self.remove_expired();
-        let query = self
-            .queries
-            .lock()
-            .unwrap()
-            .get(id)
-            .map(|entry| entry.query.clone());
+        let query = self.queries.lock().unwrap().get(&id).cloned();
         if let Some(query) = &query {
             query.touch();
         }
@@ -552,52 +517,15 @@ impl QueryRegistry {
     }
 
     fn remove_expired(&self) {
-        self.queries.lock().unwrap().retain(|_, entry| {
-            !entry
-                .query
-                .registry_expired(Arc::strong_count(&entry.query) == 1)
-        });
-    }
-
-    fn remove_oldest_terminal(&self) -> Option<OwnedSemaphorePermit> {
-        let mut queries = self.queries.lock().unwrap();
-        let oldest = queries
-            .iter()
-            .filter_map(|(id, entry)| entry.query.terminal_at.get().map(|at| (id.clone(), *at)))
-            .min_by_key(|(_, at)| *at)
-            .map(|(id, _)| id);
-        oldest
-            .and_then(|id| queries.remove(&id))
-            .map(|entry| entry._permit)
-    }
-
-    fn remove_oldest_abandoned(&self) -> Option<OwnedSemaphorePermit> {
-        let mut queries = self.queries.lock().unwrap();
-        let oldest = queries
-            .iter()
-            .filter(|(_, entry)| {
-                entry.query.terminal_at.get().is_none() && Arc::strong_count(&entry.query) == 1
-            })
-            .map(|(id, entry)| (id.clone(), *entry.query.last_accessed.lock().unwrap()))
-            .min_by_key(|(_, last_accessed)| *last_accessed)
-            .map(|(id, _)| id);
-        oldest
-            .and_then(|id| queries.remove(&id))
-            .map(|entry| entry._permit)
-    }
-
-    fn capacity_error(&self) -> Error {
-        Error::Runtime {
-            message: format!(
-                "This connection already retains {} active SQL queries",
-                self.max_capacity
-            ),
-        }
+        self.queries
+            .lock()
+            .unwrap()
+            .retain(|_, query| !query.registry_expired(Arc::strong_count(query) == 1));
     }
 }
 
 struct RemoteQuery {
-    id: String,
+    id: Uuid,
     client: Arc<SqlClientInner>,
     default_namespace_path: Vec<String>,
     state: Mutex<PollInfo>,
@@ -623,7 +551,7 @@ enum QueryLifecycle {
 
 impl RemoteQuery {
     fn new(
-        id: String,
+        id: Uuid,
         client: Arc<SqlClientInner>,
         default_namespace_path: Vec<String>,
         poll_info: PollInfo,
@@ -981,14 +909,14 @@ impl RemoteQuery {
         self.touch();
         if self.is_cancellation_requested() {
             let state = self.state.lock().await.clone();
-            return query_description(&self.id, &state, self.lifecycle());
+            return query_description(self.id, &state, self.lifecycle());
         }
         let state = self.state.lock().await.clone();
         let state = if let Some(descriptor) = state.flight_descriptor.clone() {
             let poll_guard = tokio::select! {
                 biased;
                 _ = self.wait_for_cancellation() => return query_description(
-                    &self.id,
+                    self.id,
                     &state,
                     self.lifecycle(),
                 ),
@@ -998,7 +926,7 @@ impl RemoteQuery {
                 ) => poll_guard,
             };
             let Ok(_poll_guard) = poll_guard else {
-                return query_description(&self.id, &state, self.lifecycle());
+                return query_description(self.id, &state, self.lifecycle());
             };
             let latest = self.state.lock().await.clone();
             if latest.flight_descriptor.as_ref() != Some(&descriptor) {
@@ -1007,7 +935,7 @@ impl RemoteQuery {
                 let updated = tokio::select! {
                     biased;
                     _ = self.wait_for_cancellation() => return query_description(
-                        &self.id,
+                        self.id,
                         &latest,
                         self.lifecycle(),
                     ),
@@ -1016,7 +944,7 @@ impl RemoteQuery {
                         &self.default_namespace_path,
                     ) => match result {
                         Err(_) if self.is_cancellation_requested() => return query_description(
-                            &self.id,
+                            self.id,
                             &latest,
                             self.lifecycle(),
                         ),
@@ -1032,7 +960,7 @@ impl RemoteQuery {
         } else {
             state
         };
-        query_description(&self.id, &state, self.lifecycle())
+        query_description(self.id, &state, self.lifecycle())
     }
 
     async fn cancel(&self) -> Result<()> {
@@ -1156,9 +1084,9 @@ impl RemoteQueryHandle {
 
 #[async_trait::async_trait]
 impl QueryHandle for RemoteQueryHandle {
-    fn id(&self) -> &str {
+    fn id(&self) -> Uuid {
         self.query.touch();
-        &self.query.id
+        self.query.id
     }
 
     async fn describe(&self) -> Result<QueryDescription> {
@@ -1209,22 +1137,23 @@ impl QueryHandle for RemoteQueryHandle {
 }
 
 fn query_description(
-    id: &str,
+    id: Uuid,
     poll_info: &PollInfo,
     lifecycle: QueryLifecycle,
 ) -> Result<QueryDescription> {
     let expires_at = query_expiration(poll_info)?;
     Ok(QueryDescription {
-        id: id.to_string(),
+        id,
         status: match lifecycle {
-            QueryLifecycle::Cancelling => "cancelling",
-            QueryLifecycle::Cancelled => "cancelled",
-            QueryLifecycle::Running if poll_info.flight_descriptor.is_some() => "running",
-            QueryLifecycle::Running | QueryLifecycle::Ready | QueryLifecycle::Completed => {
-                "finished"
+            QueryLifecycle::Cancelling => QueryStatus::Cancelling,
+            QueryLifecycle::Cancelled => QueryStatus::Cancelled,
+            QueryLifecycle::Running if poll_info.flight_descriptor.is_some() => {
+                QueryStatus::Running
             }
-        }
-        .to_string(),
+            QueryLifecycle::Running | QueryLifecycle::Ready | QueryLifecycle::Completed => {
+                QueryStatus::Finished
+            }
+        },
         progress: poll_info.progress,
         expires_at,
     })
@@ -1426,12 +1355,15 @@ fn tls_config(config: Option<&TlsConfig>) -> Result<ClientTlsConfig> {
 
 fn client_with_headers(
     client: FlightServiceClient<Channel>,
-    headers: &HashMap<String, String>,
+    headers: &HeaderMap,
 ) -> Result<FlightClient> {
     let mut client = FlightClient::new_from_inner(client);
     for (key, value) in headers {
+        let value = value.to_str().map_err(|err| Error::InvalidInput {
+            message: format!("Invalid SQL metadata value for {key:?}: {err}"),
+        })?;
         client
-            .add_header(key, value)
+            .add_header(key.as_str(), value)
             .map_err(|err| Error::InvalidInput {
                 message: format!("Invalid SQL metadata header {key:?}: {err}"),
             })?;
@@ -1439,33 +1371,21 @@ fn client_with_headers(
     Ok(client)
 }
 
-fn merge_headers(
-    destination: &mut HashMap<String, String>,
-    source: &HashMap<String, String>,
-) -> Result<()> {
+fn merge_headers(destination: &mut HeaderMap, source: &HashMap<String, String>) -> Result<()> {
     for (key, value) in source {
         insert_header(destination, key, value)?;
     }
     Ok(())
 }
 
-fn insert_header(headers: &mut HashMap<String, String>, key: &str, value: &str) -> Result<()> {
-    let key = key.to_ascii_lowercase();
-    let valid_key = !key.is_empty()
-        && key.bytes().all(|byte| {
-            byte.is_ascii_lowercase() || byte.is_ascii_digit() || b"-_.".contains(&byte)
-        });
-    if !valid_key {
-        return Err(Error::InvalidInput {
-            message: format!("Invalid SQL metadata key: {key:?}"),
-        });
-    }
-    if !value.is_ascii() || value.bytes().any(|byte| !(0x20..=0x7e).contains(&byte)) {
-        return Err(Error::InvalidInput {
-            message: format!("SQL metadata must be printable ASCII: {key:?}"),
-        });
-    }
-    headers.insert(key, value.to_string());
+fn insert_header(headers: &mut HeaderMap, key: &str, value: &str) -> Result<()> {
+    let key = HeaderName::from_bytes(key.as_bytes()).map_err(|err| Error::InvalidInput {
+        message: format!("Invalid SQL metadata key {key:?}: {err}"),
+    })?;
+    let value = HeaderValue::try_from(value).map_err(|err| Error::InvalidInput {
+        message: format!("Invalid SQL metadata value for {key:?}: {err}"),
+    })?;
+    headers.insert(key, value);
     Ok(())
 }
 
@@ -1547,1134 +1467,5 @@ fn sql_error(request_id: &str, error: impl std::fmt::Display) -> Error {
 }
 
 #[cfg(test)]
-mod tests {
-    use std::sync::atomic::AtomicUsize;
-
-    use arrow_array::builder::StringDictionaryBuilder;
-    use arrow_array::{Array, Int64Array, StringArray, types::Int32Type};
-    use arrow_flight::encode::FlightDataEncoderBuilder;
-    use arrow_flight::flight_service_server::{FlightService, FlightServiceServer};
-    use arrow_flight::sql::{Any, CommandStatementQuery};
-    use arrow_flight::{
-        Action, ActionType, CancelFlightInfoResult, Criteria, Empty, FlightData, FlightEndpoint,
-        FlightInfo, HandshakeRequest, HandshakeResponse, PollInfo, PutResult, SchemaResult, Ticket,
-    };
-    use arrow_schema::{DataType, Field, Schema};
-    use futures::stream::BoxStream;
-    use futures::{StreamExt, TryStreamExt};
-    use tonic::{Request, Response, Status, Streaming};
-
-    use super::*;
-    use crate::remote::client::HeaderProvider;
-
-    #[derive(Debug, Default)]
-    struct DelayedHeaderProvider {
-        delay_next: AtomicBool,
-    }
-
-    #[async_trait::async_trait]
-    impl HeaderProvider for DelayedHeaderProvider {
-        async fn get_headers(&self) -> Result<HashMap<String, String>> {
-            if self.delay_next.swap(false, Ordering::SeqCst) {
-                tokio::time::sleep(Duration::from_millis(1_100)).await;
-            }
-            Ok(HashMap::new())
-        }
-    }
-
-    fn assert_overall_timeout<T>(result: Result<T>, operation: &str) {
-        match result {
-            Err(Error::Runtime { message }) => {
-                assert_eq!(message, format!("SQL query {operation} timed out"));
-            }
-            _ => panic!("SQL query {operation} did not honor the overall timeout"),
-        }
-    }
-
-    async fn collect_result(query: &Query) -> Result<Vec<RecordBatch>> {
-        query.reader().await?.try_collect().await
-    }
-
-    #[derive(Debug)]
-    struct CapturedHeaders {
-        database: String,
-        namespace_path: String,
-        request_id: String,
-        api_key: String,
-        database_prefix: String,
-    }
-
-    #[derive(Clone)]
-    struct TestSqlService {
-        query_count: Arc<AtomicUsize>,
-        do_get_count: Arc<AtomicUsize>,
-        cancel_count: Arc<AtomicUsize>,
-        cancel_denied_count: Arc<AtomicUsize>,
-        cancel_timeout_count: Arc<AtomicUsize>,
-        cancel_unspecified_count: Arc<AtomicUsize>,
-        cancelling_response_count: Arc<AtomicUsize>,
-        incremental_finished: Arc<AtomicBool>,
-        first_continuation_count: Arc<AtomicUsize>,
-        transient_poll_failures: Arc<AtomicUsize>,
-        headers: Arc<std::sync::Mutex<Vec<CapturedHeaders>>>,
-        result: RecordBatch,
-        large_result: RecordBatch,
-        dictionary_result: RecordBatch,
-    }
-
-    impl Default for TestSqlService {
-        fn default() -> Self {
-            let schema = Arc::new(Schema::new(vec![Field::new(
-                "value",
-                DataType::Int64,
-                false,
-            )]));
-            let result =
-                RecordBatch::try_new(schema, vec![Arc::new(Int64Array::from(vec![42_i64]))])
-                    .unwrap();
-            let large_schema = Arc::new(Schema::new(vec![Field::new(
-                "value",
-                DataType::Utf8,
-                false,
-            )]));
-            let large_result = RecordBatch::try_new(
-                large_schema,
-                vec![Arc::new(StringArray::from(vec![
-                    "x".repeat(5 * 1024 * 1024),
-                ]))],
-            )
-            .unwrap();
-            let mut dictionary_builder = StringDictionaryBuilder::<Int32Type>::new();
-            dictionary_builder.append("dictionary value").unwrap();
-            let dictionary = dictionary_builder.finish();
-            let dictionary_schema = Arc::new(Schema::new(vec![Field::new(
-                "value",
-                dictionary.data_type().clone(),
-                false,
-            )]));
-            let dictionary_result =
-                RecordBatch::try_new(dictionary_schema, vec![Arc::new(dictionary)]).unwrap();
-            Self {
-                query_count: Arc::new(AtomicUsize::new(0)),
-                do_get_count: Arc::new(AtomicUsize::new(0)),
-                cancel_count: Arc::new(AtomicUsize::new(0)),
-                cancel_denied_count: Arc::new(AtomicUsize::new(0)),
-                cancel_timeout_count: Arc::new(AtomicUsize::new(0)),
-                cancel_unspecified_count: Arc::new(AtomicUsize::new(0)),
-                cancelling_response_count: Arc::new(AtomicUsize::new(0)),
-                incremental_finished: Arc::new(AtomicBool::new(false)),
-                first_continuation_count: Arc::new(AtomicUsize::new(0)),
-                transient_poll_failures: Arc::new(AtomicUsize::new(0)),
-                headers: Arc::new(std::sync::Mutex::new(Vec::new())),
-                result,
-                large_result,
-                dictionary_result,
-            }
-        }
-    }
-
-    #[tonic::async_trait]
-    impl FlightService for TestSqlService {
-        type HandshakeStream = BoxStream<'static, std::result::Result<HandshakeResponse, Status>>;
-        type ListFlightsStream = BoxStream<'static, std::result::Result<FlightInfo, Status>>;
-        type DoGetStream = BoxStream<'static, std::result::Result<FlightData, Status>>;
-        type DoPutStream = BoxStream<'static, std::result::Result<PutResult, Status>>;
-        type DoActionStream = BoxStream<'static, std::result::Result<arrow_flight::Result, Status>>;
-        type ListActionsStream = BoxStream<'static, std::result::Result<ActionType, Status>>;
-        type DoExchangeStream = BoxStream<'static, std::result::Result<FlightData, Status>>;
-
-        async fn handshake(
-            &self,
-            _request: Request<Streaming<HandshakeRequest>>,
-        ) -> std::result::Result<Response<Self::HandshakeStream>, Status> {
-            Err(Status::unimplemented("handshake"))
-        }
-
-        async fn list_flights(
-            &self,
-            _request: Request<Criteria>,
-        ) -> std::result::Result<Response<Self::ListFlightsStream>, Status> {
-            Err(Status::unimplemented("list_flights"))
-        }
-
-        async fn get_flight_info(
-            &self,
-            _request: Request<FlightDescriptor>,
-        ) -> std::result::Result<Response<FlightInfo>, Status> {
-            Err(Status::unimplemented("get_flight_info"))
-        }
-
-        async fn poll_flight_info(
-            &self,
-            request: Request<arrow_flight::FlightDescriptor>,
-        ) -> std::result::Result<Response<PollInfo>, Status> {
-            let metadata = request.metadata();
-            let header = |name| {
-                metadata
-                    .get(name)
-                    .and_then(|value| value.to_str().ok())
-                    .unwrap()
-                    .to_string()
-            };
-            self.headers.lock().unwrap().push(CapturedHeaders {
-                database: header("database"),
-                namespace_path: header("namespace-path"),
-                request_id: header("x-request-id"),
-                api_key: header("x-api-key"),
-                database_prefix: header("x-lancedb-database-prefix"),
-            });
-
-            let command = Any::decode(request.get_ref().cmd.as_ref())
-                .ok()
-                .and_then(|any| any.unpack::<CommandStatementQuery>().ok().flatten());
-            let (query, stage) = if let Some(command) = command {
-                self.query_count.fetch_add(1, Ordering::SeqCst);
-                (command.query, 0_u8)
-            } else {
-                let continuation = std::str::from_utf8(request.get_ref().cmd.as_ref())
-                    .map_err(|_| Status::invalid_argument("invalid continuation"))?;
-                let mut parts = continuation.splitn(3, ':');
-                if parts.next() != Some("poll") {
-                    return Err(Status::invalid_argument("invalid continuation"));
-                }
-                let stage = parts
-                    .next()
-                    .and_then(|stage| stage.parse().ok())
-                    .ok_or_else(|| Status::invalid_argument("invalid continuation"))?;
-                if stage == 1 {
-                    self.first_continuation_count.fetch_add(1, Ordering::SeqCst);
-                }
-                let query = parts
-                    .next()
-                    .ok_or_else(|| Status::invalid_argument("invalid continuation"))?;
-                (query.to_string(), stage)
-            };
-            if (query == "SELECT slow" || query == "SELECT cancelling") && stage > 0 {
-                tokio::time::sleep(Duration::from_millis(250)).await;
-            }
-            if query == "SELECT no info" && stage == 1 {
-                tokio::time::sleep(Duration::from_millis(100)).await;
-            }
-            if query == "SELECT incremental" && stage == 1 {
-                tokio::time::sleep(Duration::from_millis(250)).await;
-                self.incremental_finished.store(true, Ordering::SeqCst);
-            }
-            if stage == 1
-                && (query == "SELECT fail"
-                    || (query == "SELECT retry"
-                        && self.transient_poll_failures.fetch_add(1, Ordering::SeqCst) == 0))
-            {
-                return Err(Status::unavailable("transient polling failure"));
-            }
-            let complete = if query == "SELECT no info" {
-                stage >= 2
-            } else {
-                stage >= 1
-            };
-
-            let first_ticket = if query == "SELECT incremental" {
-                format!("{query}:first")
-            } else {
-                query.clone()
-            };
-            let mut info = FlightInfo::new().with_endpoint(
-                FlightEndpoint::new()
-                    .with_ticket(Ticket::new(first_ticket))
-                    .with_location("grpc://127.0.0.1:1"),
-            );
-            if query == "SELECT incremental" && stage > 0 {
-                info = info.with_endpoint(
-                    FlightEndpoint::new()
-                        .with_ticket(Ticket::new(format!("{query}:second")))
-                        .with_location("grpc://127.0.0.1:1"),
-                );
-            }
-            if query != "SELECT empty" {
-                let schema = if query == "SELECT large message" {
-                    self.large_result.schema_ref()
-                } else if query == "SELECT dictionary" {
-                    self.dictionary_result.schema_ref()
-                } else {
-                    self.result.schema_ref()
-                };
-                info = info.try_with_schema(schema).unwrap();
-            }
-            Ok(Response::new(PollInfo {
-                info: (query != "SELECT no info" || stage > 0).then_some(info),
-                flight_descriptor: (!complete)
-                    .then(|| FlightDescriptor::new_cmd(format!("poll:{}:{query}", stage + 1))),
-                progress: Some(if complete { 1.0 } else { 0.25 }),
-                expiration_time: None,
-            }))
-        }
-
-        async fn get_schema(
-            &self,
-            _request: Request<FlightDescriptor>,
-        ) -> std::result::Result<Response<SchemaResult>, Status> {
-            Err(Status::unimplemented("get_schema"))
-        }
-
-        async fn do_get(
-            &self,
-            request: Request<Ticket>,
-        ) -> std::result::Result<Response<<Self as FlightService>::DoGetStream>, Status> {
-            self.do_get_count.fetch_add(1, Ordering::SeqCst);
-            let ticket = request.get_ref().ticket.as_ref();
-            let empty = ticket == b"SELECT empty";
-            let slow = ticket == b"SELECT slow get";
-            let large = ticket == b"SELECT large message";
-            let result = if large {
-                self.large_result.clone()
-            } else if ticket == b"SELECT dictionary" {
-                self.dictionary_result.clone()
-            } else {
-                self.result.clone()
-            };
-            let schema = result.schema();
-            let input = futures::stream::once(async move {
-                if slow {
-                    tokio::time::sleep(Duration::from_millis(250)).await;
-                }
-                (!empty).then_some(Ok(result))
-            })
-            .filter_map(futures::future::ready);
-            let mut encoder = FlightDataEncoderBuilder::new().with_schema(schema);
-            if large {
-                encoder = encoder.with_max_flight_data_size(8 * 1024 * 1024);
-            }
-            let stream = encoder.build(input).map_err(Status::from);
-            Ok(Response::new(Box::pin(stream)))
-        }
-
-        async fn do_put(
-            &self,
-            _request: Request<Streaming<FlightData>>,
-        ) -> std::result::Result<Response<Self::DoPutStream>, Status> {
-            Err(Status::unimplemented("do_put"))
-        }
-
-        async fn do_action(
-            &self,
-            request: Request<Action>,
-        ) -> std::result::Result<Response<Self::DoActionStream>, Status> {
-            if request.get_ref().r#type != "CancelFlightInfo" {
-                return Err(Status::invalid_argument("unexpected action"));
-            }
-            self.cancel_count.fetch_add(1, Ordering::SeqCst);
-            let cancel_request = CancelFlightInfoRequest::decode(request.get_ref().body.clone())
-                .map_err(|_| Status::invalid_argument("invalid cancellation request"))?;
-            let query = cancel_request
-                .info
-                .and_then(|info| info.endpoint.into_iter().next())
-                .and_then(|endpoint| endpoint.ticket)
-                .and_then(|ticket| String::from_utf8(ticket.ticket.to_vec()).ok())
-                .ok_or_else(|| Status::invalid_argument("cancellation request had no ticket"))?;
-            if query == "SELECT cancel race" {
-                tokio::time::sleep(Duration::from_millis(250)).await;
-            }
-            if query == "SELECT cancel timeout" {
-                if self.cancel_timeout_count.fetch_add(1, Ordering::SeqCst) == 0 {
-                    tokio::time::sleep(Duration::from_millis(250)).await;
-                } else {
-                    return Err(Status::not_found("query cancellation completed"));
-                }
-            }
-            if query == "SELECT cancel missing" {
-                return Err(Status::not_found("query was not found"));
-            }
-            if query == "SELECT cancel denied" {
-                if self.cancel_denied_count.fetch_add(1, Ordering::SeqCst) == 0 {
-                    return Err(Status::permission_denied("cancellation is not allowed"));
-                }
-                return Err(Status::not_found("query was not found"));
-            }
-            if query == "SELECT cancel unspecified"
-                && self.cancel_unspecified_count.fetch_add(1, Ordering::SeqCst) > 0
-            {
-                return Err(Status::not_found("query cancellation completed"));
-            }
-            let status = if query == "SELECT cancel unspecified" {
-                CancelStatus::Unspecified
-            } else if query == "SELECT cancelling" {
-                if self
-                    .cancelling_response_count
-                    .fetch_add(1, Ordering::SeqCst)
-                    == 0
-                {
-                    CancelStatus::Cancelling
-                } else {
-                    return Err(Status::not_found("query cancellation completed"));
-                }
-            } else if query == "SELECT cancel race" {
-                CancelStatus::NotCancellable
-            } else {
-                CancelStatus::Cancelled
-            };
-            let response = arrow_flight::Result {
-                body: CancelFlightInfoResult::new(status).encode_to_vec().into(),
-            };
-            Ok(Response::new(Box::pin(futures::stream::iter([Ok(
-                response,
-            )]))))
-        }
-
-        async fn list_actions(
-            &self,
-            _request: Request<Empty>,
-        ) -> std::result::Result<Response<Self::ListActionsStream>, Status> {
-            Err(Status::unimplemented("list_actions"))
-        }
-
-        async fn do_exchange(
-            &self,
-            _request: Request<Streaming<FlightData>>,
-        ) -> std::result::Result<Response<Self::DoExchangeStream>, Status> {
-            Err(Status::unimplemented("do_exchange"))
-        }
-    }
-
-    #[tokio::test]
-    async fn submits_polls_fetches_cancels_and_reuses_client() {
-        let listener = std::net::TcpListener::bind("127.0.0.1:0").unwrap();
-        let address = listener.local_addr().unwrap();
-        drop(listener);
-
-        let service = TestSqlService::default();
-        let query_count = service.query_count.clone();
-        let do_get_count = service.do_get_count.clone();
-        let cancel_count = service.cancel_count.clone();
-        let incremental_finished = service.incremental_finished.clone();
-        let first_continuation_count = service.first_continuation_count.clone();
-        let headers = service.headers.clone();
-        let expected = service.result.clone();
-        let (shutdown_tx, shutdown_rx) = tokio::sync::oneshot::channel();
-        let server = tokio::spawn(
-            tonic::transport::Server::builder()
-                .add_service(FlightServiceServer::new(service))
-                .serve_with_shutdown(address, async {
-                    let _ = shutdown_rx.await;
-                }),
-        );
-        let mut ready = false;
-        for _ in 0..100 {
-            if tokio::net::TcpStream::connect(address).await.is_ok() {
-                ready = true;
-                break;
-            }
-            tokio::time::sleep(Duration::from_millis(10)).await;
-        }
-        assert!(ready, "SQL test server did not start");
-
-        let mut client_config = ClientConfig::default();
-        client_config.retry_config.read_retries = Some(1);
-        client_config.retry_config.backoff_factor = Some(0.0);
-        client_config.retry_config.backoff_jitter = Some(0.0);
-        client_config
-            .extra_headers
-            .insert("x-static-secret".to_string(), "static-secret".to_string());
-        let header_provider = Arc::new(DelayedHeaderProvider::default());
-        client_config.header_provider = Some(header_provider.clone());
-        let client = SqlClient::new(
-            "analytics".to_string(),
-            Some("tenant/production".to_string()),
-            "test-key".to_string(),
-            None,
-            Some(format!("grpc://{address}")),
-            client_config,
-        );
-        assert_eq!(client.initialized_client_count().await, 0);
-        assert!(!format!("{client:?}").contains("test-key"));
-        assert!(!format!("{client:?}").contains("static-secret"));
-
-        let mut timeout_client_config = ClientConfig::default();
-        timeout_client_config.timeout_config.timeout = Some(Duration::from_millis(50));
-        let timeout_header_provider = Arc::new(DelayedHeaderProvider::default());
-        timeout_client_config.header_provider = Some(timeout_header_provider.clone());
-        let timeout_client = SqlClient::new(
-            "analytics".to_string(),
-            Some("tenant/production".to_string()),
-            "test-key".to_string(),
-            None,
-            Some(format!("grpc://{address}")),
-            timeout_client_config,
-        );
-        timeout_header_provider
-            .delay_next
-            .store(true, Ordering::SeqCst);
-        assert_overall_timeout(
-            timeout_client
-                .submit("SELECT overall timeout", &["public".to_string()])
-                .await,
-            "submission",
-        );
-        let timeout_query = timeout_client
-            .submit("SELECT overall timeout", &["public".to_string()])
-            .await
-            .unwrap();
-        timeout_header_provider
-            .delay_next
-            .store(true, Ordering::SeqCst);
-        assert_overall_timeout(
-            timeout_client.describe(timeout_query.id()).await,
-            "description",
-        );
-        timeout_header_provider
-            .delay_next
-            .store(true, Ordering::SeqCst);
-        assert_overall_timeout(collect_result(&timeout_query).await, "result");
-        timeout_header_provider
-            .delay_next
-            .store(true, Ordering::SeqCst);
-        assert_overall_timeout(timeout_query.cancel().await, "cancellation");
-        timeout_query.cancel().await.unwrap();
-
-        let pre_dispatch_timeout = timeout_client
-            .submit("SELECT cancel missing", &["public".to_string()])
-            .await
-            .unwrap();
-        timeout_header_provider
-            .delay_next
-            .store(true, Ordering::SeqCst);
-        assert_overall_timeout(pre_dispatch_timeout.cancel().await, "cancellation");
-        assert!(pre_dispatch_timeout.cancel().await.is_err());
-        assert_ne!(
-            pre_dispatch_timeout.describe().await.unwrap().status,
-            "cancelled"
-        );
-
-        let rejected_cancel = timeout_client
-            .submit("SELECT cancel denied", &["public".to_string()])
-            .await
-            .unwrap();
-        assert!(rejected_cancel.cancel().await.is_err());
-        assert!(rejected_cancel.cancel().await.is_err());
-        assert_ne!(
-            rejected_cancel.describe().await.unwrap().status,
-            "cancelled"
-        );
-
-        let unspecified_cancel = timeout_client
-            .submit("SELECT cancel unspecified", &["public".to_string()])
-            .await
-            .unwrap();
-        assert!(unspecified_cancel.cancel().await.is_err());
-        assert!(unspecified_cancel.cancel().await.is_err());
-        assert_ne!(
-            unspecified_cancel.describe().await.unwrap().status,
-            "cancelled"
-        );
-        assert_eq!(
-            collect_result(&unspecified_cancel).await.unwrap(),
-            vec![expected.clone()]
-        );
-
-        let uncertain_cancel = timeout_client
-            .submit("SELECT cancel timeout", &["public".to_string()])
-            .await
-            .unwrap();
-        assert_overall_timeout(uncertain_cancel.cancel().await, "cancellation");
-        assert!(uncertain_cancel.cancel().await.is_err());
-        assert_ne!(
-            uncertain_cancel.describe().await.unwrap().status,
-            "cancelled"
-        );
-        assert_eq!(
-            collect_result(&uncertain_cancel).await.unwrap(),
-            vec![expected.clone()]
-        );
-
-        let first = client
-            .submit("SELECT 'super-secret'", &["public".to_string()])
-            .await
-            .unwrap();
-        let id_suffix = first.id().strip_prefix(QUERY_ID_PREFIX).unwrap();
-        assert_eq!(id_suffix.len(), 32);
-        assert!(id_suffix.bytes().all(|byte| byte.is_ascii_hexdigit()));
-        assert!(!first.id().contains("super-secret"));
-        header_provider.delay_next.store(true, Ordering::SeqCst);
-        let describe_started = Instant::now();
-        let first_description = client.describe(first.id()).await.unwrap();
-        assert!(describe_started.elapsed() >= Duration::from_millis(1_100));
-        assert_eq!(first_description.status, "finished");
-        assert_eq!(first_description.progress, Some(1.0));
-        let first_result = collect_result(&first).await.unwrap();
-        assert!(first.reader().await.is_err());
-
-        let incremental = client
-            .submit("SELECT incremental", &["public".to_string()])
-            .await
-            .unwrap();
-        let mut incremental_result = incremental.reader().await.unwrap();
-        let first_incremental_batch =
-            tokio::time::timeout(Duration::from_millis(100), incremental_result.try_next())
-                .await
-                .expect("the first partial result must arrive before query completion")
-                .unwrap()
-                .unwrap();
-        assert_eq!(first_incremental_batch, expected);
-        assert!(!incremental_finished.load(Ordering::SeqCst));
-        let remaining_incremental_batches =
-            incremental_result.try_collect::<Vec<_>>().await.unwrap();
-        assert_eq!(remaining_incremental_batches, vec![expected.clone()]);
-        assert!(incremental_finished.load(Ordering::SeqCst));
-
-        let interrupted_result = Arc::new(
-            client
-                .submit("SELECT no info", &["public".to_string()])
-                .await
-                .unwrap(),
-        );
-        let continuation_count_before = first_continuation_count.load(Ordering::SeqCst);
-        let interrupted_result_task = {
-            let interrupted_result = interrupted_result.clone();
-            tokio::spawn(async move { interrupted_result.reader().await })
-        };
-        tokio::time::timeout(Duration::from_millis(100), async {
-            while first_continuation_count.load(Ordering::SeqCst) == continuation_count_before {
-                tokio::task::yield_now().await;
-            }
-        })
-        .await
-        .expect("result preparation must start polling");
-        interrupted_result_task.abort();
-        assert!(
-            interrupted_result_task
-                .await
-                .is_err_and(|error| error.is_cancelled())
-        );
-        assert_eq!(
-            collect_result(&interrupted_result).await.unwrap(),
-            vec![expected.clone()],
-            "cancelling result preparation must release the one-shot result claim",
-        );
-
-        let dropped_reader = client
-            .submit("SELECT slow", &["public".to_string()])
-            .await
-            .unwrap();
-        let tracked_dropped_reader = client.queries.get(dropped_reader.id()).unwrap();
-        let continuation_count_before = first_continuation_count.load(Ordering::SeqCst);
-        let dropped_result_stream = dropped_reader.reader().await.unwrap();
-        tokio::time::timeout(Duration::from_millis(100), async {
-            while first_continuation_count.load(Ordering::SeqCst) == continuation_count_before {
-                tokio::task::yield_now().await;
-            }
-        })
-        .await
-        .expect("the result producer must start continuation polling");
-        assert!(Arc::strong_count(&tracked_dropped_reader) >= 4);
-        drop(dropped_result_stream);
-        tokio::time::timeout(Duration::from_millis(100), async {
-            while Arc::strong_count(&tracked_dropped_reader) != 3 {
-                tokio::task::yield_now().await;
-            }
-        })
-        .await
-        .expect("dropping a result reader must stop its producer");
-
-        let staged = client
-            .submit("SELECT no info", &["public".to_string()])
-            .await
-            .unwrap();
-        let staged_running = client.describe(staged.id()).await.unwrap();
-        assert_eq!(staged_running.status, "running");
-        let staged_finished = client.describe(staged.id()).await.unwrap();
-        assert_eq!(staged_finished.status, "finished");
-
-        let empty = client
-            .submit("SELECT empty", &["public".to_string()])
-            .await
-            .unwrap();
-        let empty_result = empty.reader().await.unwrap();
-        assert_eq!(empty_result.schema(), expected.schema());
-        let empty_result = empty_result.try_collect::<Vec<_>>().await.unwrap();
-
-        let large = client
-            .submit("SELECT large message", &["public".to_string()])
-            .await
-            .unwrap();
-        let large_result = collect_result(&large).await.unwrap();
-        assert_eq!(large_result.len(), 1);
-        assert_eq!(large_result[0].num_rows(), 1);
-        assert_eq!(
-            large_result[0]
-                .column(0)
-                .as_any()
-                .downcast_ref::<StringArray>()
-                .unwrap()
-                .value(0)
-                .len(),
-            5 * 1024 * 1024,
-        );
-
-        let dictionary = client
-            .submit("SELECT dictionary", &["public".to_string()])
-            .await
-            .unwrap();
-        let dictionary_result = collect_result(&dictionary).await.unwrap();
-        assert_eq!(dictionary_result.len(), 1);
-        assert_eq!(
-            dictionary_result[0].schema().field(0).data_type(),
-            &DataType::Utf8,
-        );
-        assert_eq!(
-            dictionary_result[0]
-                .column(0)
-                .as_any()
-                .downcast_ref::<StringArray>()
-                .unwrap()
-                .value(0),
-            "dictionary value",
-        );
-
-        let cancelled = client
-            .submit(
-                "SELECT cancelled",
-                &["events".to_string(), "raw".to_string()],
-            )
-            .await
-            .unwrap();
-        cancelled.cancel().await.unwrap();
-        assert_eq!(cancelled.describe().await.unwrap().status, "cancelled");
-        assert!(matches!(
-            cancelled.reader().await,
-            Err(Error::JobCancelled { .. })
-        ));
-
-        let slow = Arc::new(
-            client
-                .submit("SELECT slow", &["public".to_string()])
-                .await
-                .unwrap(),
-        );
-        let result_task = {
-            let slow = slow.clone();
-            tokio::spawn(async move { collect_result(&slow).await })
-        };
-        tokio::time::sleep(Duration::from_millis(25)).await;
-        tokio::time::timeout(Duration::from_millis(150), slow.cancel())
-            .await
-            .expect("cancellation must not wait for result polling")
-            .unwrap();
-        assert!(matches!(
-            tokio::time::timeout(Duration::from_millis(150), result_task)
-                .await
-                .expect("cancellation must wake result polling")
-                .unwrap(),
-            Err(Error::JobCancelled { .. })
-        ));
-        let cancel_count_after_slow = cancel_count.load(Ordering::SeqCst);
-        slow.cancel().await.unwrap();
-        assert_eq!(
-            cancel_count.load(Ordering::SeqCst),
-            cancel_count_after_slow,
-            "a confirmed cancellation must not be sent again",
-        );
-
-        let slow_get = Arc::new(
-            client
-                .submit("SELECT slow get", &["public".to_string()])
-                .await
-                .unwrap(),
-        );
-        let do_get_count_before_slow = do_get_count.load(Ordering::SeqCst);
-        let slow_get_result_task = {
-            let slow_get = slow_get.clone();
-            tokio::spawn(async move { collect_result(&slow_get).await })
-        };
-        while do_get_count.load(Ordering::SeqCst) == do_get_count_before_slow {
-            tokio::task::yield_now().await;
-        }
-        slow_get.cancel().await.unwrap();
-        assert_eq!(slow_get.describe().await.unwrap().status, "cancelled");
-        assert!(matches!(
-            tokio::time::timeout(Duration::from_millis(150), slow_get_result_task)
-                .await
-                .expect("cancellation must wake result fetching")
-                .unwrap(),
-            Err(Error::JobCancelled { .. })
-        ));
-        assert!(slow_get.reader().await.is_err());
-
-        let restored = Arc::new(
-            RemoteQuery::new(
-                "restored".to_string(),
-                client.inner.clone(),
-                vec!["public".to_string()],
-                PollInfo {
-                    flight_descriptor: Some(FlightDescriptor::new_cmd("restored")),
-                    ..Default::default()
-                },
-            )
-            .unwrap(),
-        );
-        let mut restored_waiter = {
-            let restored = restored.clone();
-            tokio::spawn(async move { restored.wait_for_cancellation().await })
-        };
-        tokio::task::yield_now().await;
-        restored.mark_cancelling();
-        restored.restore_after_rejected_cancellation().await;
-        assert_eq!(restored.lifecycle(), QueryLifecycle::Running);
-        assert!(
-            tokio::time::timeout(Duration::from_millis(25), &mut restored_waiter)
-                .await
-                .is_err(),
-            "a stale cancellation notification must not complete the waiter",
-        );
-        restored.mark_cancelling();
-        tokio::time::timeout(Duration::from_millis(100), restored_waiter)
-            .await
-            .expect("a current cancellation must complete the waiter")
-            .unwrap();
-
-        let cancelling = Arc::new(
-            client
-                .submit("SELECT cancelling", &["public".to_string()])
-                .await
-                .unwrap(),
-        );
-        let cancelling_result_task = {
-            let cancelling = cancelling.clone();
-            tokio::spawn(async move { collect_result(&cancelling).await })
-        };
-        tokio::time::sleep(Duration::from_millis(25)).await;
-        cancelling.cancel().await.unwrap();
-        assert_eq!(cancelling.describe().await.unwrap().status, "cancelling");
-        assert!(matches!(
-            tokio::time::timeout(Duration::from_millis(150), cancelling_result_task)
-                .await
-                .expect("an accepted cancellation must wake result polling")
-                .unwrap(),
-            Err(Error::JobCancelled { .. })
-        ));
-        cancelling.cancel().await.unwrap();
-        assert_eq!(cancelling.describe().await.unwrap().status, "cancelled");
-
-        let cancel_race = Arc::new(
-            client
-                .submit("SELECT cancel race", &["public".to_string()])
-                .await
-                .unwrap(),
-        );
-        let cancel_count_before_race = cancel_count.load(Ordering::SeqCst);
-        let cancel_race_task = {
-            let cancel_race = cancel_race.clone();
-            tokio::spawn(async move { cancel_race.cancel().await })
-        };
-        while cancel_count.load(Ordering::SeqCst) == cancel_count_before_race {
-            tokio::task::yield_now().await;
-        }
-        let cancel_race_result = collect_result(&cancel_race).await.unwrap();
-        tokio::time::timeout(Duration::from_millis(500), cancel_race_task)
-            .await
-            .expect("completed result must make in-flight cancellation a no-op")
-            .unwrap()
-            .unwrap();
-        assert_eq!(cancel_race.describe().await.unwrap().status, "finished");
-        assert_eq!(cancel_race_result, vec![expected.clone()]);
-        assert!(cancel_race.reader().await.is_err());
-
-        let no_info = Arc::new(
-            client
-                .submit("SELECT no info", &["public".to_string()])
-                .await
-                .unwrap(),
-        );
-        let continuation_count_before = first_continuation_count.load(Ordering::SeqCst);
-        let no_info_result_task = {
-            let no_info = no_info.clone();
-            tokio::spawn(async move { collect_result(&no_info).await })
-        };
-        tokio::time::sleep(Duration::from_millis(10)).await;
-        tokio::time::timeout(Duration::from_secs(1), no_info.cancel())
-            .await
-            .expect("cancellation should wait for cancellable query information")
-            .unwrap();
-        assert!(matches!(
-            no_info_result_task.await.unwrap(),
-            Err(Error::JobCancelled { .. })
-        ));
-        assert_eq!(
-            first_continuation_count.load(Ordering::SeqCst),
-            continuation_count_before + 1,
-            "result and cancel must share one continuation poll",
-        );
-
-        let retried = client
-            .submit("SELECT retry", &["public".to_string()])
-            .await
-            .unwrap();
-        assert_eq!(
-            collect_result(&retried).await.unwrap(),
-            vec![expected.clone()]
-        );
-
-        let failed = client
-            .submit("SELECT fail", &["public".to_string()])
-            .await
-            .unwrap();
-        assert!(collect_result(&failed).await.is_err());
-
-        let active_registry = QueryRegistry::new(1);
-        let active_permit = active_registry.reserve().unwrap();
-        let active_query = Arc::new(
-            RemoteQuery::new(
-                "active".to_string(),
-                client.inner.clone(),
-                vec!["public".to_string()],
-                PollInfo {
-                    flight_descriptor: Some(FlightDescriptor::new_cmd("active")),
-                    ..Default::default()
-                },
-            )
-            .unwrap(),
-        );
-        active_registry.insert("active".to_string(), active_query.clone(), active_permit);
-        assert!(active_registry.reserve().is_err());
-        assert!(Arc::ptr_eq(
-            &active_registry.get("active").unwrap(),
-            &active_query,
-        ));
-
-        let expired_registry = QueryRegistry::new(1);
-        let expired_permit = expired_registry.reserve().unwrap();
-        let expired_query = Arc::new(
-            RemoteQuery::new(
-                "expired".to_string(),
-                client.inner.clone(),
-                vec!["public".to_string()],
-                PollInfo {
-                    flight_descriptor: Some(FlightDescriptor::new_cmd("expired")),
-                    expiration_time: Some(Default::default()),
-                    ..Default::default()
-                },
-            )
-            .unwrap(),
-        );
-        expired_registry.insert("expired".to_string(), expired_query, expired_permit);
-        assert!(expired_registry.reserve().is_ok());
-        assert!(expired_registry.get("expired").is_none());
-
-        let terminal_registry = QueryRegistry::new(1);
-        let terminal_permit = terminal_registry.reserve().unwrap();
-        let terminal_query = Arc::new(
-            RemoteQuery::new(
-                "terminal".to_string(),
-                client.inner.clone(),
-                vec!["public".to_string()],
-                PollInfo {
-                    flight_descriptor: None,
-                    expiration_time: Some(Default::default()),
-                    ..Default::default()
-                },
-            )
-            .unwrap(),
-        );
-        terminal_registry.insert("terminal".to_string(), terminal_query, terminal_permit);
-        assert!(terminal_registry.get("terminal").is_some());
-        assert!(terminal_registry.reserve().is_ok());
-        assert!(terminal_registry.get("terminal").is_none());
-
-        let transitioned_registry = QueryRegistry::new(1);
-        let transitioned_permit = transitioned_registry.reserve().unwrap();
-        let transition_descriptor = FlightDescriptor::new_cmd("transitioned");
-        let transitioned_query = Arc::new(
-            RemoteQuery::new(
-                "transitioned".to_string(),
-                client.inner.clone(),
-                vec!["public".to_string()],
-                PollInfo {
-                    flight_descriptor: Some(transition_descriptor.clone()),
-                    ..Default::default()
-                },
-            )
-            .unwrap(),
-        );
-        transitioned_query
-            .update_state(
-                &transition_descriptor,
-                PollInfo {
-                    flight_descriptor: None,
-                    expiration_time: Some(Default::default()),
-                    ..Default::default()
-                },
-            )
-            .await
-            .unwrap();
-        transitioned_registry.insert(
-            "transitioned".to_string(),
-            transitioned_query,
-            transitioned_permit,
-        );
-        assert!(transitioned_registry.get("transitioned").is_some());
-
-        let abandoned_registry = QueryRegistry::new(1);
-        let abandoned_permit = abandoned_registry.reserve().unwrap();
-        let abandoned_query = Arc::new(
-            RemoteQuery::new(
-                "abandoned".to_string(),
-                client.inner.clone(),
-                vec!["public".to_string()],
-                PollInfo {
-                    flight_descriptor: Some(FlightDescriptor::new_cmd("abandoned")),
-                    ..Default::default()
-                },
-            )
-            .unwrap(),
-        );
-        abandoned_registry.insert(
-            "abandoned".to_string(),
-            abandoned_query.clone(),
-            abandoned_permit,
-        );
-        drop(abandoned_query);
-        assert!(abandoned_registry.reserve().is_ok());
-        assert!(abandoned_registry.get("abandoned").is_none());
-
-        let stale_registry = QueryRegistry::new(1);
-        let stale_permit = stale_registry.reserve().unwrap();
-        let stale_query = Arc::new(
-            RemoteQuery::new(
-                "stale".to_string(),
-                client.inner.clone(),
-                vec!["public".to_string()],
-                PollInfo {
-                    flight_descriptor: Some(FlightDescriptor::new_cmd("stale")),
-                    ..Default::default()
-                },
-            )
-            .unwrap(),
-        );
-        *stale_query.last_accessed.lock().unwrap() = Instant::now() - ABANDONED_QUERY_RETENTION;
-        stale_registry.insert("stale".to_string(), stale_query.clone(), stale_permit);
-        drop(stale_query);
-        assert!(stale_registry.get("stale").is_none());
-
-        let concurrent_registry = Arc::new(QueryRegistry::new(2));
-        for id in ["terminal-one", "terminal-two"] {
-            let permit = concurrent_registry.reserve().unwrap();
-            let query = Arc::new(
-                RemoteQuery::new(
-                    id.to_string(),
-                    client.inner.clone(),
-                    vec!["public".to_string()],
-                    PollInfo::default(),
-                )
-                .unwrap(),
-            );
-            concurrent_registry.insert(id.to_string(), query, permit);
-        }
-        let reservation_barrier = Arc::new(std::sync::Barrier::new(3));
-        let reserve = |registry: Arc<QueryRegistry>, barrier: Arc<std::sync::Barrier>| {
-            tokio::task::spawn_blocking(move || {
-                barrier.wait();
-                registry.reserve()
-            })
-        };
-        let reservation_one = reserve(concurrent_registry.clone(), reservation_barrier.clone());
-        let reservation_two = reserve(concurrent_registry.clone(), reservation_barrier.clone());
-        reservation_barrier.wait();
-        let (permit_one, permit_two) = tokio::join!(reservation_one, reservation_two);
-        let permit_one = permit_one.unwrap().unwrap();
-        let permit_two = permit_two.unwrap().unwrap();
-        assert_eq!(concurrent_registry.capacity.available_permits(), 0);
-        drop((permit_one, permit_two));
-        assert_eq!(concurrent_registry.capacity.available_permits(), 2);
-
-        assert_eq!(client.initialized_client_count().await, 1);
-        assert_eq!(query_count.load(Ordering::SeqCst), 21);
-        assert_eq!(do_get_count.load(Ordering::SeqCst), 17);
-        assert_eq!(cancel_count.load(Ordering::SeqCst), 15);
-        assert_eq!(first_result, vec![expected.clone()]);
-        assert!(empty_result.is_empty());
-        assert!(client.describe("invalid").await.is_err());
-        {
-            let headers = headers.lock().unwrap();
-            assert_eq!(headers[0].database, "analytics");
-            assert_eq!(headers[0].namespace_path, "public");
-            assert_eq!(headers[0].api_key, "test-key");
-            assert_eq!(headers[0].database_prefix, "tenant/production");
-            assert!(
-                headers
-                    .iter()
-                    .any(|header| header.namespace_path == "events$raw")
-            );
-            assert!(
-                headers
-                    .windows(2)
-                    .all(|headers| headers[0].request_id != headers[1].request_id)
-            );
-        }
-        let _ = shutdown_tx.send(());
-        server.await.unwrap().unwrap();
-    }
-
-    #[test]
-    fn normalizes_supported_uris() {
-        assert_eq!(
-            normalize_sql_host_override("grpc://localhost").unwrap(),
-            SqlTarget {
-                uri: "http://localhost:10025".to_string(),
-                tls: false,
-            }
-        );
-        assert_eq!(
-            normalize_sql_host_override("grpcs://example.com").unwrap(),
-            SqlTarget {
-                uri: "https://example.com:10026".to_string(),
-                tls: true,
-            }
-        );
-        assert_eq!(
-            normalize_sql_host_override("grpc://[::1]:10025").unwrap(),
-            SqlTarget {
-                uri: "http://[::1]:10025".to_string(),
-                tls: false,
-            }
-        );
-        assert_eq!(
-            normalize_sql_host_override("https://example.com:443").unwrap(),
-            SqlTarget {
-                uri: "https://example.com:443".to_string(),
-                tls: true,
-            }
-        );
-    }
-
-    #[test]
-    fn derives_plaintext_endpoint_from_host_override() {
-        assert_eq!(
-            resolve_sql_host_override(Some("http://localhost:10024"), None).unwrap(),
-            SqlTarget {
-                uri: "http://localhost:10025".to_string(),
-                tls: false,
-            }
-        );
-        assert_eq!(
-            resolve_sql_host_override(Some("http://localhost:80"), None).unwrap(),
-            SqlTarget {
-                uri: "http://localhost:81".to_string(),
-                tls: false,
-            }
-        );
-    }
-
-    #[test]
-    fn rejects_unsafe_or_ambiguous_endpoints() {
-        assert!(normalize_sql_host_override("ftp://localhost").is_err());
-        assert!(normalize_sql_host_override("grpc://user@localhost").is_err());
-        assert!(normalize_sql_host_override("grpc://localhost/path").is_err());
-        assert!(resolve_sql_host_override(Some("https://localhost"), None).is_err());
-    }
-
-    #[test]
-    fn validates_namespace_components() {
-        assert!(validate_namespace_path(&[]).is_ok());
-        assert!(validate_namespace_path(&["events".into(), "raw".into()]).is_ok());
-        assert!(validate_namespace_path(&["events$raw".into()]).is_err());
-        assert!(validate_namespace_path(&["".into()]).is_err());
-        assert!(validate_namespace_path(&["café".into()]).is_err());
-    }
-}
+#[path = "sql_test.rs"]
+mod tests;
